Retorna a composição da comissão

Retorna dataframe com os dados dos membros de uma comissão

fetch_composicao_comissao(sigla, casa, orgaos_camara)

Arguments

sigla
sigla da comissão

Value

Dataframe com os dados dos membros de uma comissão

Examples

fetch_composicao_comissao("CCJ",'senado')

#> [1] "Baixando composição da comissão CCJ em senado"

#>              cargo id   partido uf situacao                    nome
#> 1             <NA>  1       MDB AM  Titular           Eduardo Braga
#> 2             <NA>  1       MDB AL Suplente         Renan Calheiros
#> 3       PRESIDENTE  1       MDB MS  Titular            Simone Tebet
#> 4             <NA>  1       MDB PE Suplente Fernando Bezerra Coelho
#> 5             <NA>  1       PRB RR  Titular         Mecias de Jesus
#> 6             <NA>  1       MDB AC Suplente           Marcio Bittar
#> 7             <NA>  1       MDB PA  Titular          Jader Barbalho
#> 8             <NA>  1       MDB PI Suplente          Marcelo Castro
#> 9             <NA>  1       MDB PB  Titular           José Maranhão
#> 10            <NA>  1       MDB SC Suplente            Dário Berger
#> 11            <NA>  1        PP PI  Titular           Ciro Nogueira
#> 12            <NA>  1        PP PB Suplente        Daniella Ribeiro
#> 13            <NA>  1        PP SC  Titular         Esperidião Amin
#> 14            <NA>  1        PP RS Suplente      Luis Carlos Heinze
#> 15            <NA>  2      PSDB MG  Titular       Antonio Anastasia
#> 16            <NA>  2      PSDB SP Suplente              José Serra
#> 17            <NA>  2      PSDB CE  Titular        Tasso Jereissati
#> 18            <NA>  2      PSDB MA Suplente           Roberto Rocha
#> 19            <NA>  2      PODE PI  Titular           Elmano Férrer
#> 20            <NA>  2      PSDB AL Suplente           Rodrigo Cunha
#> 21            <NA>  2      PODE PR  Titular     Oriovisto Guimarães
#> 22            <NA>  2      PODE RS Suplente          Lasier Martins
#> 23            <NA>  2      PODE ES  Titular         Rose de Freitas
#> 24            <NA>  2       PSL SP Suplente           Major Olimpio
#> 25            <NA>  2       PSL MT  Titular             Juíza Selma
#> 26            <NA>  2       PSL RJ Suplente        Flávio Bolsonaro
#> 27            <NA>  3       PSB PB  Titular Veneziano Vital do Rêgo
#> 28            <NA>  3       PSB GO Suplente            Jorge Kajuru
#> 29            <NA>  3       PDT CE  Titular               Cid Gomes
#> 30            <NA>  3 CIDADANIA ES Suplente           Marcos do Val
#> 31            <NA>  3      REDE ES  Titular       Fabiano Contarato
#> 32            <NA>  3      REDE AP Suplente      Randolfe Rodrigues
#> 33            <NA>  3 CIDADANIA SE  Titular       Alessandro Vieira
#> 34            <NA>  3       PDT RO Suplente            Acir Gurgacz
#> 35            <NA>  3       PDT MA  Titular                Weverton
#> 36            <NA>  3       PSB DF Suplente            Leila Barros
#> 37            <NA>  4        PT PE  Titular          Humberto Costa
#> 38            <NA>  4      PROS RR Suplente           Telmário Mota
#> 39            <NA>  4      PROS AL  Titular         Renilde Bulhões
#> 40            <NA>  4        PT BA Suplente           Jaques Wagner
#> 41            <NA>  4        PT SE  Titular        Rogério Carvalho
#> 42            <NA>  4        PT PA Suplente             Paulo Rocha
#> 43            <NA>  5       PSD BA  Titular            Otto Alencar
#> 44            <NA>  5       PSD AC Suplente          Sérgio Petecão
#> 45            <NA>  5       PSD BA  Titular          Angelo Coronel
#> 46            <NA>  5       PSD MS Suplente           Nelsinho Trad
#> 47            <NA>  5       PSD RJ  Titular      Arolde de Oliveira
#> 48            <NA>  5       PSD MG Suplente            Carlos Viana
#> 49            <NA>  6       DEM MG  Titular         Rodrigo Pacheco
#> 50            <NA>  6       PSC PA Suplente        Zequinha Marinho
#> 51            <NA>  6       DEM RO  Titular          Marcos Rogério
#> 52            <NA>  6       DEM SE Suplente    Maria do Carmo Alves
#> 53 VICE-PRESIDENTE  6        PR SC  Titular          Jorginho Mello
#> 54            <NA>  6        PR MT Suplente     Wellington Fagundes
#>                                                                     foto sigla
#> 1  http://www.senado.leg.br/senadores/img/fotos-oficiais/senador4994.jpg   CCJ
#> 2    http://www.senado.leg.br/senadores/img/fotos-oficiais/senador70.jpg   CCJ
#> 3  http://www.senado.leg.br/senadores/img/fotos-oficiais/senador5527.jpg   CCJ
#> 4  http://www.senado.leg.br/senadores/img/fotos-oficiais/senador5540.jpg   CCJ
#> 5  http://www.senado.leg.br/senadores/img/fotos-oficiais/senador6027.jpg   CCJ
#> 6   http://www.senado.leg.br/senadores/img/fotos-oficiais/senador285.jpg   CCJ
#> 7    http://www.senado.leg.br/senadores/img/fotos-oficiais/senador35.jpg   CCJ
#> 8   http://www.senado.leg.br/senadores/img/fotos-oficiais/senador742.jpg   CCJ
#> 9  http://www.senado.leg.br/senadores/img/fotos-oficiais/senador3361.jpg   CCJ
#> 10 http://www.senado.leg.br/senadores/img/fotos-oficiais/senador5537.jpg   CCJ
#> 11  http://www.senado.leg.br/senadores/img/fotos-oficiais/senador739.jpg   CCJ
#> 12 http://www.senado.leg.br/senadores/img/fotos-oficiais/senador5998.jpg   CCJ
#> 13   http://www.senado.leg.br/senadores/img/fotos-oficiais/senador22.jpg   CCJ
#> 14 http://www.senado.leg.br/senadores/img/fotos-oficiais/senador1186.jpg   CCJ
#> 15 http://www.senado.leg.br/senadores/img/fotos-oficiais/senador5529.jpg   CCJ
#> 16   http://www.senado.leg.br/senadores/img/fotos-oficiais/senador90.jpg   CCJ
#> 17 http://www.senado.leg.br/senadores/img/fotos-oficiais/senador3396.jpg   CCJ
#> 18  http://www.senado.leg.br/senadores/img/fotos-oficiais/senador677.jpg   CCJ
#> 19 http://www.senado.leg.br/senadores/img/fotos-oficiais/senador5531.jpg   CCJ
#> 20 http://www.senado.leg.br/senadores/img/fotos-oficiais/senador5905.jpg   CCJ
#> 21 http://www.senado.leg.br/senadores/img/fotos-oficiais/senador5924.jpg   CCJ
#> 22 http://www.senado.leg.br/senadores/img/fotos-oficiais/senador5533.jpg   CCJ
#> 23 http://www.senado.leg.br/senadores/img/fotos-oficiais/senador2331.jpg   CCJ
#> 24 http://www.senado.leg.br/senadores/img/fotos-oficiais/senador5666.jpg   CCJ
#> 25 http://www.senado.leg.br/senadores/img/fotos-oficiais/senador5929.jpg   CCJ
#> 26 http://www.senado.leg.br/senadores/img/fotos-oficiais/senador5894.jpg   CCJ
#> 27 http://www.senado.leg.br/senadores/img/fotos-oficiais/senador5748.jpg   CCJ
#> 28 http://www.senado.leg.br/senadores/img/fotos-oficiais/senador5895.jpg   CCJ
#> 29 http://www.senado.leg.br/senadores/img/fotos-oficiais/senador5973.jpg   CCJ
#> 30 http://www.senado.leg.br/senadores/img/fotos-oficiais/senador5942.jpg   CCJ
#> 31 http://www.senado.leg.br/senadores/img/fotos-oficiais/senador5953.jpg   CCJ
#> 32 http://www.senado.leg.br/senadores/img/fotos-oficiais/senador5012.jpg   CCJ
#> 33 http://www.senado.leg.br/senadores/img/fotos-oficiais/senador5982.jpg   CCJ
#> 34 http://www.senado.leg.br/senadores/img/fotos-oficiais/senador4981.jpg   CCJ
#> 35 http://www.senado.leg.br/senadores/img/fotos-oficiais/senador5411.jpg   CCJ
#> 36 http://www.senado.leg.br/senadores/img/fotos-oficiais/senador5979.jpg   CCJ
#> 37 http://www.senado.leg.br/senadores/img/fotos-oficiais/senador5008.jpg   CCJ
#> 38 http://www.senado.leg.br/senadores/img/fotos-oficiais/senador5535.jpg   CCJ
#> 39 http://www.senado.leg.br/senadores/img/fotos-oficiais/senador5561.jpg   CCJ
#> 40  http://www.senado.leg.br/senadores/img/fotos-oficiais/senador581.jpg   CCJ
#> 41 http://www.senado.leg.br/senadores/img/fotos-oficiais/senador5352.jpg   CCJ
#> 42  http://www.senado.leg.br/senadores/img/fotos-oficiais/senador374.jpg   CCJ
#> 43 http://www.senado.leg.br/senadores/img/fotos-oficiais/senador5523.jpg   CCJ
#> 44 http://www.senado.leg.br/senadores/img/fotos-oficiais/senador4560.jpg   CCJ
#> 45 http://www.senado.leg.br/senadores/img/fotos-oficiais/senador5967.jpg   CCJ
#> 46 http://www.senado.leg.br/senadores/img/fotos-oficiais/senador5985.jpg   CCJ
#> 47  http://www.senado.leg.br/senadores/img/fotos-oficiais/senador751.jpg   CCJ
#> 48 http://www.senado.leg.br/senadores/img/fotos-oficiais/senador5990.jpg   CCJ
#> 49 http://www.senado.leg.br/senadores/img/fotos-oficiais/senador5732.jpg   CCJ
#> 50 http://www.senado.leg.br/senadores/img/fotos-oficiais/senador3806.jpg   CCJ
#> 51 http://www.senado.leg.br/senadores/img/fotos-oficiais/senador5422.jpg   CCJ
#> 52 http://www.senado.leg.br/senadores/img/fotos-oficiais/senador1023.jpg   CCJ
#> 53 http://www.senado.leg.br/senadores/img/fotos-oficiais/senador5350.jpg   CCJ
#> 54 http://www.senado.leg.br/senadores/img/fotos-oficiais/senador1173.jpg   CCJ
#>      casa
#> 1  senado
#> 2  senado
#> 3  senado
#> 4  senado
#> 5  senado
#> 6  senado
#> 7  senado
#> 8  senado
#> 9  senado
#> 10 senado
#> 11 senado
#> 12 senado
#> 13 senado
#> 14 senado
#> 15 senado
#> 16 senado
#> 17 senado
#> 18 senado
#> 19 senado
#> 20 senado
#> 21 senado
#> 22 senado
#> 23 senado
#> 24 senado
#> 25 senado
#> 26 senado
#> 27 senado
#> 28 senado
#> 29 senado
#> 30 senado
#> 31 senado
#> 32 senado
#> 33 senado
#> 34 senado
#> 35 senado
#> 36 senado
#> 37 senado
#> 38 senado
#> 39 senado
#> 40 senado
#> 41 senado
#> 42 senado
#> 43 senado
#> 44 senado
#> 45 senado
#> 46 senado
#> 47 senado
#> 48 senado
#> 49 senado
#> 50 senado
#> 51 senado
#> 52 senado
#> 53 senado
#> 54 senado