Recupera o histórico de relatorias de uma proposição

Retorna dataframe com o histórico de relatorias, contendo data e nome do relator. No senado contem informacoes extras sobre motivo do fim da relatoria e informacoes do relator. Na camara, contem o a sigla do local da relatoria Ao fim, a função retira todos as colunas que tenham tipo lista para uniformizar o dataframe.

get_relatorias(proposicao_id, casa, last_n = NULL)

Arguments

proposicao_id
ID de uma proposição do Senado

last_n
Pegar os ultimos N relatores

Value

Dataframe com as informações detalhadas do histórico de relatorias de uma proposição no Senado

Examples

get_relatorias(91341, 'senado', 3)

#> Warning: `cols` is now required.
#> Please use `cols = c()`

#> # A tibble: 3 x 20
#>   codigo_tipo_rel… descricao_tipo_… data_designacao data_destituicao
#>   <chr>            <chr>            <chr>           <chr>           
#> 1 1                Relator          2015-05-04      2016-05-17      
#> 2 1                Relator          2011-03-04      2014-12-22      
#> 3 1                Relator          2010-07-08      2010-12-22      
#> # … with 16 more variables: descricao_motivo_destituicao <chr>,
#> #   codigo_parlamentar <chr>, nome_parlamentar <chr>,
#> #   nome_completo_parlamentar <chr>, sexo_parlamentar <chr>,
#> #   forma_tratamento <chr>, url_foto_parlamentar <chr>,
#> #   url_pagina_parlamentar <chr>, email_parlamentar <chr>,
#> #   sigla_partido_parlamentar <chr>, uf_parlamentar <chr>,
#> #   codigo_comissao <chr>, sigla_comissao <chr>, nome_comissao <chr>,
#> #   sigla_casa_comissao <chr>, nome_casa_comissao <chr>