Processa lista de apensadas com árvore partindo das apensadas monitoradas até chegar na proposição raiz.

Recupera uma lista que contém as ligações entre as proposições apensadas monitoradas e suas proposições principais para uma casa de origem.

process_lista_apensadas_por_casa(
  proposicoes_apensadas,
  casa_origem = "camara",
  fresh_execution = FALSE,
  export_filepath = "lista_apensadas.csv",
  save_result = TRUE
)

Arguments

proposicoes_apensadas
Dataframe com lista de proposições apensadas e suas proposições principais. É necessário pelo menos 3 colunas: id_ext, casa, id_prop_principal.

casa_origem
Casa de origem das proposições. Parâmetro usado para recuperar dados da uriPrincipal.

fresh_execution
Se FALSE então um arquivo com a lista de apensadas executada anteriormente será utilizada. TRUE caso uma execução nova seja o objetivo.

export_filepath
Caminho para salvar lista de apensadas

save_result
TRUE se o resultado da lista deve ser salvo em csv no caminho do export_filepath

Value

Lista com árvore de apensados.