Retorna a composição da comissão

Retorna dataframe com os dados dos membros de uma comissão

fetch_composicao_comissao(
  sigla,
  casa,
  orgaos_camara,
  data_inicio_arg = "2019-02-01"
)

Arguments

sigla
sigla da comissão

Value

Dataframe com os dados dos membros de uma comissão

Examples

fetch_composicao_comissao("CCJ",'senado')

#> [1] "Baixando composição da comissão CCJ no(a) senado"

#> Warning: `cols` is now required.
#> Please use `cols = c(MEMBROS.MEMBROS_ROW)`

#> Warning: `cols` is now required.
#> Please use `cols = c()`

#> # A tibble: 53 x 9
#>    cargo id    partido uf    situacao nome       foto                sigla casa 
#>    <chr> <chr> <chr>   <chr> <chr>    <chr>      <chr>               <chr> <chr>
#>  1 NA    4994  MDB     AM    Titular  Eduardo B… http://www.senado.… CCJ   sena…
#>  2 NA    3777  MDB     TO    Suplente Eduardo G… http://www.senado.… CCJ   sena…
#>  3 NA    70    MDB     AL    Titular  Renan Cal… http://www.senado.… CCJ   sena…
#>  4 NA    5748  MDB     PB    Suplente Veneziano… http://www.senado.… CCJ   sena…
#>  5 NA    5527  MDB     MS    Titular  Simone Te… http://www.senado.… CCJ   sena…
#>  6 NA    285   MDB     AC    Suplente Marcio Bi… http://www.senado.… CCJ   sena…
#>  7 NA    5540  MDB     PE    Titular  Fernando … http://www.senado.… CCJ   sena…
#>  8 NA    5585  MDB     GO    Suplente Luiz do C… http://www.senado.… CCJ   sena…
#>  9 NA    2331  MDB     ES    Titular  Rose de F… http://www.senado.… CCJ   sena…
#> 10 NA    35    MDB     PA    Suplente Jader Bar… http://www.senado.… CCJ   sena…
#> # … with 43 more rows